% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getResSeq}
\alias{getResSeq}
\title{resSeq IDs}
\usage{
getResSeq(coord)
}
\arguments{
\item{coord}{coordinate object (data frame)}
}
\value{
Vector of resSeq IDs
}
\description{
Returns resSeq IDs for a coordinate object.
}
\examples{
\dontrun{
getResSeq(selectRecords(pdb, chainID = "B"))
}
}

