% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSequence}
\alias{getSequence}
\title{Get sequence}
\usage{
getSequence(coord)
}
\arguments{
\item{coord}{coordinate object (data frame)}
}
\value{
Vector of residue names
}
\description{
Returns the sequence of a coordinate object (as a vector).
}
\examples{
\dontrun{
getSequence(selectRecords(pdb, chainID = "A"))
}
}

