% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.pdb}
\alias{read.pdb}
\title{Read PDB file}
\usage{
read.pdb(filename)
}
\arguments{
\item{filename}{name of the input .pdb file}
}
\value{
pdb object (see description below)
}
\description{
Reads a .pdb file into a pdb object.
}
\section{Fields}{

\describe{
\item{\code{header}}{data frame with PDB header}

\item{\code{crystal}}{data frame with crystallographic data}

\item{\code{coord}}{data frame with coordinate data}

\item{\code{footer}}{data frame with PDB footer}
}}

