% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectRecords}
\alias{selectRecords}
\title{Select records}
\usage{
selectRecords(pdb, ...)
}
\arguments{
\item{pdb}{pdb object}

\item{...}{selection criteria (see Examples)}
}
\value{
Coordinate object (data frame) containing the selected subset
}
\description{
Returns a subset of a pdb object's coord data.
}
\examples{
\dontrun{
selectRecords(pdb, resSeq = 1:10)
selectRecords(pdb, element = "C", resSeq = c(20, 22:25))
selectRecords(pdb, name = "P", resName = c("G", "C"), chainID = "A")
}
}

