% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seqToFasta}
\alias{seqToFasta}
\title{Convert sequence to FASTA}
\usage{
seqToFasta(sequence, name = "Sequence", wrapAt = 80)
}
\arguments{
\item{sequence}{vector containing valid residue codes}

\item{name}{name for the output sequence}

\item{wrapAt}{column, at which the sequence is wrapped
(set to 0 for unwrapped output)}
}
\value{
String in FASTA format
}
\description{
Converts sequence vector to a sequence string in FASTA format
}
\examples{
\dontrun{
chainB = selectRecords(pdb, chainID = "B")
sequence = getSequence(chainB)
seqToFasta(sequence)
}
}

