% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{showStructure}
\alias{showStructure}
\title{Show structure}
\usage{
showStructure(coord, color_by = "element", color_other = "pink",
  color_table = if (color_by == "element") ELEMENT_COLOR_TABLE else
  CHAIN_COLOR_TABLE, type = "s", size = 1.2, box = FALSE, axes = FALSE,
  xlab = "", ylab = "", zlab = "", ...)
}
\arguments{
\item{coord}{coordinate object (data frame)}

\item{color_by}{name of the column used for coloring}

\item{color_other}{color associated with atoms not found in color_table}

\item{color_table}{list of colors associated with different atoms}

\item{...}{any valid parameter for \link{plot3d}}
}
\description{
Shows an interactive 3D view of the structure from given coordinates
in a graphics device (window, Web browser etc.)
}

