% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superimpose.R
\name{superimpose}
\alias{superimpose}
\title{Superimpose structures}
\usage{
superimpose(mobile, target_subset, mobile_subset)
}
\arguments{
\item{mobile}{coord data of the mobile structure}

\item{target_subset, mobile_subset}{atom subsets to be aligned
(subsets should be the same length and in the same order)}
}
\value{
Coord data of the superimposed structure
}
\description{
Superimposes the mobile structure on the target structure
using Kabsch algorithm.
}
\examples{
\dontrun{
superimpose(pdb2$coord,
            selectRecords(pdb1, name = "P", resSeq = 1:10),
            selectRecords(pdb2, name = "P", resSeq = 2:11))
}
}
\references{
Kabsch, Wolfgang, (1976), A solution for the best rotation to relate
two sets of vectors. \emph{Acta Crystallographica} \strong{32}:922.
\url{https://dx.doi.org/10.1107/S0567739476001873}
}

