% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.pdb}
\alias{write.pdb}
\title{Write PDB file}
\usage{
write.pdb(data, filename)
}
\arguments{
\item{data}{pdb object}

\item{filename}{name of the output .pdb file}
}
\description{
Writes a pdb object to a file.
}

